+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2020-10-05
 ### Added
 - Shipping fee feature.
 
@@ -46,5 +48,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release.
 
 [Unreleased]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.2.0...HEAD
+[0.3.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/my-jam-store/stripe-checkout-service/releases/tag/0.1.0